Several changes
I reduced the amount of lines on the Javascript file, now the show/hide member function is short and userful. I discarted the function that highlight the navbar items when the content is on the screen, and added a smooth scroll animation. In the end I started the member profile card, I did the struture and the initial style.