Member's card update
The member's card style and logic (appear and disappear) are finished. Some information of the members was added too, but it isn't complete. The function to control data visibility was polished, removing some lines and incresing the funcionality. I started to implement the back-to-the-top fixed button that is showed only when the player scroll down, but only the link is working. I already removed the y scrollbar.